fix: api cleanup
