--- src/apis/noop.ts
@@ -1,3 +1,0 @@
-export default (): void => {
-  /* do nothing */
-};
