+++ src/components/games/slots/ReelDisplay.tsx
@@ -4,7 +4,7 @@ import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
 import Typography from "@mui/material/Typography";
-import { SlotDisplay } from "../../../apis/SlotMachine";
+import { SlotDisplay } from "./slotMachine";
 
 interface ReelDisplayProps {
   reel: SlotDisplay[];
