+++ src/components/games/slots/slotMachine.ts
