+++ src/components/resume/comp-calc/CompEntryDisplay.tsx
@@ -7,7 +7,7 @@ import Grid from "@mui/material/Grid2";
 import Typography from "@mui/material/Typography";
 import { CompCalcEntry, CompEntry } from "../../../jotai/comp-calculator-state";
 import dateObj from "../../../apis/DateHelper";
-import usDollar from "../../../apis/usdFormatter";
+import usDollar from "../../../apis/usDollar";
 
 interface CompEntryDisplayProps {
   compEntries: CompEntry[];
